+++ protected/humhub/modules/content/live/NewContent.php
@@ -12,7 +12,7 @@ use humhub\modules\live\components\LiveEvent;
 
 /**
  * Live event for new contents
- * 
+ *
  * @since 1.2
  */
 class NewContent extends LiveEvent
@@ -22,19 +22,19 @@ class NewContent extends LiveEvent
      * @var int the id of the new content
      */
     public $contentId;
-    
+
     /**
-     * @var string space guid for space content container 
+     * @var string space guid for space content container
      */
     public $sguid;
-    
+
     /**
-     * @var string user guid for user content container 
+     * @var string user guid for user content container
      */
     public $uguid;
-    
+
     /**
-     * @var string originator guid 
+     * @var string originator guid
      */
     public $originator;
 
@@ -52,5 +52,11 @@ class NewContent extends LiveEvent
      * @var boolean if true it's meant to be a silent content creation
      */
     public $silent;
+
+    /**
+     * @var string|null content stream_channel
+     * @since 1.5
+     */
+    public $streamChannel;
 
 }