+++ protected/humhub/modules/stream/resources/js/humhub.stream.js
@@ -6,6 +6,7 @@
 humhub.module('stream', function (module, require, $) {
 
     var Component = require('action').Component;
+    var event = require('event');
 
     /**
      * Returns a stream instance for the given selector. If no selector is given this function returns the main wall stream.
@@ -23,13 +24,18 @@ humhub.module('stream', function (module, require, $) {
      * Returns a single entry with a given contentId of the wall stream
      *
      * @deprecated since v1.3 use Component.instance instead
-     * @param id
+     * @param contentId
      */
     var getEntry = function (contentId) {
         return module.getStream().entry(contentId);
     };
 
+    var unload = function() {
+        event.off('.stream');
+    };
+
     module.export({
+        unload: unload,
         getStream: getStream,
         getEntry: getEntry
     });